requirements.txt
fastapi
uvicorn
pydantic
fpdf2
